+++ apps/ui-service/src/shared/lib/format-date.ts
@@ -1,0 +1,7 @@
+import { format } from "date-fns"
+
+export function formatDate(
+  dateString: string | number | null | undefined | Date
+) {
+  return format(dateString ? new Date(dateString) : new Date(), "MMM, do yyyy")
+}